+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:example/alert_section.dart';
 import 'package:example/avatar_group_label_section.dart';
+import 'package:example/avatar_group_section.dart';
 import 'package:example/avatar_section.dart';
 import 'package:example/badge_section.dart';
 import 'package:example/button_section.dart';
@@ -34,6 +35,7 @@ class MainApp extends StatelessWidget {
               AlertSection(),
               AvatarSection(),
               AvatarGroupLabelSection(),
+              AvatarGroupSection(),
               ButtonSection(),
               IconShapeSection(),
               InputFieldSection(),